Split protocol negotiation tests into modular suites
